Poll for source load on startup
Problem: on startup the script loads before the sources do, so when the script goes to create the callback it throws an error. The function isn't called after the sources load so the callback is not set until the user changes the source or reloads the script.

Here's my brute force solution: add a timer and check every 10 milliseconds for the source we're looking for. Once it's found, ditch the timer.

This is brute force and inelegant, but it does solve the problem and the callback is properly created on program start. It would be far better to replace this with a timely signal callback, but so far I've had no luck finding one.